@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
     <PackageReference Include="NServiceBus" Version="8.0.3" />
-    <PackageReference Include="ServiceComposer.AspNetCore" Version="1.13.0" />
+    <PackageReference Include="ServiceComposer.AspNetCore" Version="2.0.0" />
   </ItemGroup>
 
   <ItemGroup>
