@@ -7,20 +7,20 @@ namespace NServiceBus
     {
         public static void ApplyCommonConfiguration(this EndpointConfiguration endpointConfiguration, Action<RoutingSettings<LearningTransport>> configureRouting = null)
         {
-            endpointConfiguration.UseSerialization<NewtonsoftSerializer>();
+            endpointConfiguration.UseSerialization<NewtonsoftJsonSerializer>();
             var transportConfig = endpointConfiguration.UseTransport<LearningTransport>();
             configureRouting?.Invoke(transportConfig.Routing());
 
             endpointConfiguration.AuditProcessedMessagesTo("audit");
             endpointConfiguration.SendFailedMessagesTo("error");
 
-            endpointConfiguration.SendHeartbeatTo(
-                serviceControlQueue: "Particular.ServiceControl",
-                frequency: TimeSpan.FromSeconds(10),
-                timeToLive: TimeSpan.FromSeconds(5));
-
-            endpointConfiguration.AuditSagaStateChanges(
-                serviceControlQueue: "Particular.ServiceControl");
+            // endpointConfiguration.SendHeartbeatTo(
+            //     serviceControlQueue: "Particular.ServiceControl",
+            //     frequency: TimeSpan.FromSeconds(10),
+            //     timeToLive: TimeSpan.FromSeconds(5));
+            //
+            // endpointConfiguration.AuditSagaStateChanges(
+            //     serviceControlQueue: "Particular.ServiceControl");
 
             var messageConventions = endpointConfiguration.Conventions();
             messageConventions.DefiningMessagesAs(t => t.Namespace != null && t.Namespace.EndsWith(".Messages"));
