@@ -3,11 +3,12 @@ using Microsoft.AspNetCore.Routing;
 using Microsoft.EntityFrameworkCore;
 using ServiceComposer.AspNetCore;
 using System.Linq;
+using Microsoft.AspNetCore.Mvc;
 using Ticketing.ViewModelComposition.Events;
 
 namespace Finance.ViewModelComposition
 {
-    class AvailableTicketsLoadedSubscriber : ISubscribeToCompositionEvents
+    class AvailableTicketsLoadedSubscriber : ICompositionEventsSubscriber
     {
         public bool Matches(RouteData routeData, string httpVerb, HttpRequest request)
         {
@@ -20,21 +21,20 @@ namespace Finance.ViewModelComposition
                    && !routeData.Values.ContainsKey("id");
         }
 
-        public void Subscribe(IPublishCompositionEvents publisher)
+        [HttpGet("/home/index")]
+        public void Subscribe(ICompositionEventsPublisher publisher)
         {
-            publisher.Subscribe<AvailableTicketsLoaded>(async (requestId, pageViewModel, @event, rd, req) =>
+            publisher.Subscribe<AvailableTicketsLoaded>(async (@event, request) =>
             {
                 var ids = @event.AvailableTicketsViewModel.Keys.ToArray();
-                using (var db = Data.FinanceContext.Create())
-                {
-                    var ticketPrices = await db.TicketPrices
-                        .Where(ticketPrice => ids.Contains(ticketPrice.Id))
-                        .ToListAsync();
+                await using var db = Data.FinanceContext.Create();
+                var ticketPrices = await db.TicketPrices
+                    .Where(ticketPrice => ids.Contains(ticketPrice.Id))
+                    .ToListAsync();
 
-                    foreach (var ticketPrice in ticketPrices)
-                    {
-                        @event.AvailableTicketsViewModel[(int)ticketPrice.Id].TicketPrice = ticketPrice.Price;
-                    }
+                foreach (var ticketPrice in ticketPrices)
+                {
+                    @event.AvailableTicketsViewModel[(int)ticketPrice.Id].TicketPrice = ticketPrice.Price;
                 }
             });
         }
