@@ -5,10 +5,11 @@ using Reservations.Messages.Commands;
 using ServiceComposer.AspNetCore;
 using System;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Reservations.ViewModelComposition
 {
-    class ReservationsReservePostHandler : IHandleRequests
+    class ReservationsReservePostHandler : ICompositionRequestsHandler
     {
         private readonly IMessageSession messageSession;
 
@@ -17,18 +18,8 @@ namespace Reservations.ViewModelComposition
             this.messageSession = messageSession;
         }
 
-        public bool Matches(RouteData routeData, string httpVerb, HttpRequest request)
-        {
-            var controller = (string)routeData.Values["controller"];
-            var action = (string)routeData.Values["action"];
-
-            return HttpMethods.IsPost(httpVerb)
-                   && controller.ToLowerInvariant() == "reservations"
-                   && action.ToLowerInvariant() == "reserve"
-                   && routeData.Values.ContainsKey("id");
-        }
-
-        public Task Handle(string requestId, dynamic vm, RouteData routeData, HttpRequest request)
+        [HttpPost("/reservations/reserve/{id}")]
+        public Task Handle(HttpRequest request)
         {
             /*
              * In a production environment if multiple services are interested in the
@@ -39,12 +30,12 @@ namespace Reservations.ViewModelComposition
              * production environment. In order to make this part safe, which is not the
              * scope of this demo asynchronous messaging should be introduced earlier in
              * the processing pipeline.
-             * 
+             *
              * More information: https://milestone.topics.it/2019/05/02/safety-first.html
              */
             var message = new ReserveTicket()
             {
-                TicketId = int.Parse((string)routeData.Values["id"]),
+                TicketId = int.Parse((string)request.HttpContext.GetRouteValue("id")),
                 ReservationId = new Guid(request.Cookies["reservation-id"])
             };
 
