@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Routing;
 using Microsoft.EntityFrameworkCore;
 using Reservations.ViewModelComposition.Events;
 using ServiceComposer.AspNetCore;
@@ -8,24 +7,18 @@ using System.Collections.Generic;
 using System.Dynamic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Reservations.ViewModelComposition
 {
-    class TicketsReservationGetHandler : IHandleRequests
+    class TicketsReservationGetHandler : ICompositionRequestsHandler
     {
-        public bool Matches(RouteData routeData, string httpVerb, HttpRequest request)
+        [HttpGet("/reservations/index")]
+        [HttpGet("/reservations/review")]
+        public async Task Handle(HttpRequest request)
         {
-            var controller = (string)routeData.Values["controller"];
-            var action = (string)routeData.Values["action"];
+            var vm = request.GetComposedResponseModel();
 
-            return HttpMethods.IsGet(httpVerb)
-                   && controller.ToLowerInvariant() == "reservations"
-                   && (action.ToLowerInvariant() == "index" || action.ToLowerInvariant() == "review")
-                   && !routeData.Values.ContainsKey("id");
-        }
-
-        public async Task Handle(string requestId, dynamic vm, RouteData routeData, HttpRequest request)
-        {
             if (!request.Cookies.ContainsKey("reservation-id"))
             {
                 vm.Reservation = null;
@@ -33,49 +26,48 @@ namespace Reservations.ViewModelComposition
             }
 
             var reservationId = new Guid(request.Cookies["reservation-id"]);
-            using (var db = Data.ReservationsContext.Create())
+            await using var db = Data.ReservationsContext.Create();
+            var reservation = await db.Reservations
+                .Where(r => r.Id == reservationId)
+                .Include(r => r.ReservedTickets)
+                .SingleOrDefaultAsync();
+
+            if (reservation == null)
             {
-                var reservation = await db.Reservations
-                    .Where(r => r.Id == reservationId)
-                    .Include(r => r.ReservedTickets)
-                    .SingleOrDefaultAsync();
-
-                if (reservation == null)
-                {
-                    vm.Reservation = null;
-                    return;
-                }
-
-                vm.Reservation = new ExpandoObject();
-                vm.Reservation.Id = reservationId;
-
-                IEnumerable<(int TicketId, int Quantity)> reservedTickeks = reservation.ReservedTickets
-                    .GroupBy(t => t.TicketId)
-                    .Select(g => (g.Key, g.Count()));
-
-                var reservedTicketsViewModel = MapToDictionary(reservedTickeks);
-
-                await vm.RaiseEvent(new ReservedTicketsLoaded()
-                {
-                    Reservation = vm.Reservation, 
-                    ReservedTicketsViewModel = reservedTicketsViewModel
-                });
-
-                vm.Reservation.ReservedTickets = reservedTicketsViewModel.Values.ToList();
+                vm.Reservation = null;
+                return;
             }
+
+            vm.Reservation = new ExpandoObject();
+            vm.Reservation.Id = reservationId;
+
+            IEnumerable<(int TicketId, int Quantity)> reservedTickets = reservation.ReservedTickets
+                .GroupBy(t => t.TicketId)
+                .Select(g => (g.Key, g.Count()));
+
+            var reservedTicketsViewModel = MapToDictionary(reservedTickets);
+
+            var compositionContext = request.GetCompositionContext();
+            await compositionContext.RaiseEvent(new ReservedTicketsLoaded()
+            {
+                Reservation = vm.Reservation,
+                ReservedTicketsViewModel = reservedTicketsViewModel
+            });
+
+            vm.Reservation.ReservedTickets = reservedTicketsViewModel.Values.ToList();
         }
 
-        IDictionary<int, dynamic> MapToDictionary(IEnumerable<(int TicketId, int Quantity)> reservedTickets)
+        static IDictionary<int, dynamic> MapToDictionary(IEnumerable<(int TicketId, int Quantity)> reservedTickets)
         {
             var reservedTicketsViewModel = new Dictionary<int, dynamic>();
 
-            foreach (var reservedTicket in reservedTickets)
+            foreach ((int ticketId, int quantity) in reservedTickets)
             {
                 dynamic vm = new ExpandoObject();
-                vm.TicketId = reservedTicket.TicketId;
-                vm.Quantity = reservedTicket.Quantity;
+                vm.TicketId = ticketId;
+                vm.Quantity = quantity;
 
-                reservedTicketsViewModel[reservedTicket.TicketId] = vm;
+                reservedTicketsViewModel[ticketId] = vm;
             }
 
             return reservedTicketsViewModel;
