@@ -1,15 +1,15 @@
 ﻿using Shipping.Messages.Commands;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Routing;
 using NServiceBus;
 using ServiceComposer.AspNetCore;
 using System;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 using Shipping.Data;
 
 namespace Shipping.ViewModelComposition
 {
-    class ReservationsCheckoutPostHandler : IHandleRequests
+    class ReservationsCheckoutPostHandler : ICompositionRequestsHandler
     {
         private readonly IMessageSession messageSession;
 
@@ -18,17 +18,8 @@ namespace Shipping.ViewModelComposition
             this.messageSession = messageSession;
         }
 
-        public bool Matches(RouteData routeData, string httpVerb, HttpRequest request)
-        {
-            var controller = (string)routeData.Values["controller"];
-            var action = (string)routeData.Values["action"];
-
-            return HttpMethods.IsPost(httpVerb)
-                   && controller.ToLowerInvariant() == "reservations"
-                   && action.ToLowerInvariant() == "checkout";
-        }
-
-        public Task Handle(string requestId, dynamic vm, RouteData routeData, HttpRequest request)
+        [HttpPost("/reservations/checkout")]
+        public Task Handle(HttpRequest request)
         {
             /*
              * In a production environment if multiple services are interested in the
