@@ -2,22 +2,14 @@
 using Microsoft.AspNetCore.Routing;
 using ServiceComposer.AspNetCore;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Shipping.ViewModelComposition
 {
-    class ReservationsFinalizePostHandler : IHandleRequests
+    class ReservationsFinalizePostHandler : ICompositionRequestsHandler
     {
-        public bool Matches(RouteData routeData, string httpVerb, HttpRequest request)
-        {
-            var controller = (string)routeData.Values["controller"];
-            var action = (string)routeData.Values["action"];
-
-            return HttpMethods.IsPost(httpVerb)
-                   && controller.ToLowerInvariant() == "reservations"
-                   && action.ToLowerInvariant() == "finalize";
-        }
-
-        public Task Handle(string requestId, dynamic vm, RouteData routeData, HttpRequest request)
+        [HttpPost("/reservations/finalize")]
+        public Task Handle(HttpRequest request)
         {
             var response = request.HttpContext.Response;
             response.Cookies.Append("reservation-delivery-option-id", request.Form["DeliveryOption"]);
